requirements.txt
aiodns==3.5.0
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
fonttools==4.59.0
frozenlist==1.7.0
idna==3.10
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.3
multidict==6.6.3
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
yarl==1.20.1
